refactor: rename hash.rs to felt.rs
